2020/1_vr_station/server/requirements.txt
asgiref==3.2.3
Django==3.0.2
django-cors-headers==3.2.1
django-rest-framework==0.1.0
djangorestframework==3.11.0
Pillow==8.1.2
pytz==2019.3
sqlparse==0.3.0
